+++ src/main/java/com/akatsuki/newsum/domain/webtoon/repository/CommentQueryRepositoryImpl.java
@@ -8,8 +8,8 @@ import java.util.List;
 import com.akatsuki.newsum.common.pagination.model.cursor.Cursor;
 import com.akatsuki.newsum.common.pagination.model.query.QueryEngineType;
 import com.akatsuki.newsum.common.pagination.model.query.QueryFragment;
-import com.akatsuki.newsum.common.pagination.querybuilder.CursorPageQueryBuilder;
-import com.akatsuki.newsum.common.pagination.querybuilder.registry.CursorPageQueryRegistry;
+import com.akatsuki.newsum.common.pagination.query.CursorPageQueryBuilder;
+import com.akatsuki.newsum.common.pagination.query.registry.CursorPageQueryRegistry;
 import com.akatsuki.newsum.domain.webtoon.dto.CommentReadDto;
 import com.akatsuki.newsum.domain.webtoon.dto.QCommentReadDto;
 import com.akatsuki.newsum.domain.webtoon.entity.comment.entity.QComment;
