+++ src/main/java/com/akatsuki/newsum/domain/webtoon/repository/WebtoonQueryRepositoryImpl.java
@@ -15,8 +15,8 @@ import com.akatsuki.newsum.common.pagination.model.cursor.CreatedAtIdCursor;
 import com.akatsuki.newsum.common.pagination.model.cursor.Cursor;
 import com.akatsuki.newsum.common.pagination.model.query.QueryEngineType;
 import com.akatsuki.newsum.common.pagination.model.query.QueryFragment;
-import com.akatsuki.newsum.common.pagination.querybuilder.CursorPageQueryBuilder;
-import com.akatsuki.newsum.common.pagination.querybuilder.registry.CursorPageQueryRegistry;
+import com.akatsuki.newsum.common.pagination.query.CursorPageQueryBuilder;
+import com.akatsuki.newsum.common.pagination.query.registry.CursorPageQueryRegistry;
 import com.akatsuki.newsum.domain.webtoon.entity.webtoon.Category;
 import com.akatsuki.newsum.domain.webtoon.entity.webtoon.QWebtoon;
 import com.akatsuki.newsum.domain.webtoon.entity.webtoon.RecentView;
